requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.4.0
PyJWT==2.8.0
Flask-CORS==4.0.0